requirements.txt
fastapi==0.105.0
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
pydantic==2.5.2
gunicorn==21.2.0